Move Login into its own .jsx file.
